backend/requirements.txt
# LeadRadar Backend Dependencies
# Python 3.9+

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP & Scraping
httpx>=0.26.0
beautifulsoup4>=4.12.0
duckduckgo-search>=4.0.0

# Database
supabase>=2.3.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
